+++ tidecv/__init__.py
@@ -1,5 +1,5 @@
-__version__ = '1.0.1'
+__version__ = '1.1'
 
-from .quantify import *
+from .tide import *
 from .errors.qualifiers import *
 from . import datasets