+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # TIDE Changelog
 
+## v1.1
+### Better Data Output
+ - Added better output and visualization of data.
+ - Its now possible to compare multiple predictions of different models to the same Groundtruth file
+
 ## v1.0.1
 ### Error Calculation
  - **Important**: Fixed an oversight where detections ignored by AP calculation were not allowed to contribute to fixing errors. This caused a lot of error in datasets with large amounts of ignore regions (e.g., LVIS) to significantly overrepresent Missed Error and underrepresent either Classification or Localization error. This fix will also slightly change errors for other datasets (< .4 dAP on COCO), but conclusions on LVIS change dramatically.